Add sorting and searching to challenges index by creating a separate sorting controller
